renaming graph to rdf_to_graph extractor
