+++ cognite/neat/graph/extractors/__init__.py
@@ -1,4 +1,4 @@
-from cognite.neat.graph.extractors import graph
+from cognite.neat.graph.extractors import rdf_to_graph
 from cognite.neat.graph.extractors.graph_sheet_to_graph import (
     extract_graph_from_sheet,
     read_graph_excel_file_to_table_by_name,
@@ -6,7 +6,7 @@ from cognite.neat.graph.extractors.graph_sheet_to_graph import (
 from cognite.neat.graph.extractors.graph_store import NeatGraphStore, drop_graph_store
 
 __all__ = [
-    "graph",
+    "rdf_to_graph",
     "NeatGraphStore",
     "extract_graph_from_sheet",
     "read_graph_excel_file_to_table_by_name",
